support S3 STAC ingest
